update unit test screenshot and separate AlertTest and dataTest
