+++ src/test/java/data_management_Test/FileDataReaderTest.java
@@ -1,5 +1,9 @@
-package com.data_management;
+package data_management_Test;
 
+import com.data_management.DataStorage;
+import com.data_management.FileDataReader;
+import com.data_management.Patient;
+import com.data_management.PatientRecord;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
